Enhanced code and added comments
