+++ .github/workflows/homepage_json_gen_cron_job.yml
@@ -37,7 +37,7 @@ jobs:
          pip install -r requirements.txt
 
     - name: Execute Python script
-      run: python generate_homepage_xml.py
+      run: python generate_homepage_json.py
 
     - name: Configure Git
       run: |
